+++ libs/cdk/src/lib/components/pagination/utils/pagination.util.ts
@@ -6,8 +6,10 @@ export const paginate = (value?: PaginateOptions | null) => {
     return null;
   }
 
+  const { omitFirstLast = false, omitPreviousNext = false, pagesBeforeAfter = 2, firstPage = 1 } = value;
+
   const pages: PaginationItem[] = [];
-  const activePage = clamp(value.currentPage, 1, value.totalPageCount);
+  const activePage = clamp(value.currentPage, firstPage, value.totalPageCount);
 
   if (value.totalPageCount === 0) {
     return null;
@@ -27,28 +29,46 @@ export const paginate = (value?: PaginateOptions | null) => {
     return url.toString();
   };
 
-  pages.push({
-    page: 1,
-    current: false,
-    ariaLabel: 'First page',
-    disabled: activePage === 1,
-    type: 'hotLink',
-    explicitType: 'first',
-    url: createUrl(1),
-  });
-  pages.push({
-    page: activePage - 1,
-    current: false,
-    ariaLabel: `Previous page`,
-    disabled: activePage === 1,
-    type: 'hotLink',
-    explicitType: 'previous',
-    url: createUrl(activePage - 1),
-  });
+  if (!omitFirstLast) {
+    pages.push({
+      page: firstPage,
+      current: false,
+      ariaLabel: 'First page',
+      disabled: activePage === firstPage,
+      type: 'hotLink',
+      explicitType: 'first',
+      url: createUrl(firstPage),
+    });
+  }
 
-  // add 2 pages before and after active page
-  for (let i = activePage - 2; i <= activePage + 2; i++) {
-    if (i > 0 && i <= value.totalPageCount) {
+  if (!omitPreviousNext) {
+    pages.push({
+      page: activePage - 1,
+      current: false,
+      ariaLabel: `Previous page`,
+      disabled: activePage === 1,
+      type: 'hotLink',
+      explicitType: 'previous',
+      url: createUrl(activePage - 1),
+    });
+  }
+
+  let pagesBefore = activePage - pagesBeforeAfter;
+  let pagesAfter = activePage + pagesBeforeAfter;
+
+  if (pagesBefore < firstPage) {
+    pagesAfter += firstPage - pagesBefore;
+    pagesBefore = firstPage;
+  }
+
+  if (pagesAfter > value.totalPageCount) {
+    pagesBefore -= pagesAfter - value.totalPageCount;
+    pagesAfter = value.totalPageCount;
+  }
+
+  // add N pages before and after active page
+  for (let i = pagesBefore; i <= pagesAfter; i++) {
+    if (i > firstPage - 1 && i <= value.totalPageCount) {
       const explicitType =
         i === activePage
           ? 'current'
@@ -68,24 +88,29 @@ export const paginate = (value?: PaginateOptions | null) => {
     }
   }
 
-  pages.push({
-    page: activePage + 1,
-    current: false,
-    ariaLabel: `Next page`,
-    disabled: activePage === value.totalPageCount,
-    type: 'hotLink',
-    explicitType: 'next',
-    url: createUrl(activePage + 1),
-  });
-  pages.push({
-    page: value.totalPageCount,
-    current: false,
-    ariaLabel: `Last page`,
-    disabled: activePage === value.totalPageCount,
-    type: 'hotLink',
-    explicitType: 'last',
-    url: createUrl(value.totalPageCount),
-  });
+  if (!omitPreviousNext) {
+    pages.push({
+      page: activePage + 1,
+      current: false,
+      ariaLabel: `Next page`,
+      disabled: activePage === value.totalPageCount,
+      type: 'hotLink',
+      explicitType: 'next',
+      url: createUrl(activePage + 1),
+    });
+  }
+
+  if (!omitFirstLast) {
+    pages.push({
+      page: value.totalPageCount,
+      current: false,
+      ariaLabel: `Last page`,
+      disabled: activePage === value.totalPageCount,
+      type: 'hotLink',
+      explicitType: 'last',
+      url: createUrl(value.totalPageCount),
+    });
+  }
 
   return pages;
 };